Add support for obfuscated resource IDs.
Most of Sagittariidae's resources are externally referencable
resouruces.  We depend on the HashId algorithm (http://hashids.org) to
generate unique IDs that don't carry unintened meaning (e.g. relative
ordering) or leak implementation-specific detail.

These identifiers are generated from an internal identifier (a
per-resource integer-value that is incremented as new resources are
added) and a salt.  The salt prevents the same hash IDs being generated
for the different resources with the same internal ID.  Datomic's
SqUUIDs could have served, but the integer ID and salt are necessary to
maintain compatibility with the first incarnation of Sagittariidae that
was used to capture data: the Python prototype.

Both integer IDs and the obfuscated IDs (Hashids) are allocated in a
transaction function when a new resource is added.  The transaction
function guarantees serialisation and atomicity, but has the
disadvantage of forcing us to add only a single resource of any type per
transaction.  This is because all invocations of the transaction
function will see the same value of the data (i.e. the pre-transaction
value), and as a result, all invocations of the function will allocate
the *same* numeric ID.

For now, our anticipated throughput is low enough that it this
restriction won't cause a bottleneck.