Functions to add projects, methods and samples.
The functions that construct transaction data now return a complete
transaction list, rather than only a single element for the entity being
constructed.  This is needed to handle the scenario where the addition
of an entity requires multiple statements in the TX data.  For example,
when adding a sample, the sample entity is both created and added to the
project to which it belongs.

Note also that sample names are namespaced by the project to guarantee
uniqueness.  In a relational DB, this would be a UNQUE index over and
the sample name and a backref to the project ID.