Refactor `liaison` tests.
The superfluous Vars have been been removed in favour of direct
references to the private fns in the namespace.  Using the Vars saved
only 4 chars and added an unecessary level of indirection.

`find-by-name` has been made more specific to retrieve only the
`obfuscated-id` for a resource, as that is the pattern when referring to
resources.