Make `get-sample` multi-arity.
When retriving a sample by ID, the project ID isn't actually necessary.
It is only present as a safety net to ensure, when it is used in a
URI that includes a project ID, that the request is coherent (i.e that
the sample does exist within the project).

But since the sample IDs are globally unique, internally we should be
able to retrieve a sample using only the ID.

Brandon Bloom's [backtick] library has been added to smooth the construction of
the query by allowing the familiar unquote syntax to be used.

[backtick]: https://github.com/brandonbloom/backtick