Least-effort FE-compatible search.
As we continue to build toward feature parity with the Python prototype,
it would be helpful to be able to drive the WS using the existing FE.
Since stage and datafile management depend on being able to retrieve
some subset of the samples for a project, we need to enable a sort of
search.

This least-effort change simply returns all samples for a project.