Retrieve the stages of sample testing.
Stages are returned as an ordered list, since the tests performed in any
given stage may irrevocably alter the properties of that sample; it is
important that we return the sequence in the same order that the tests
were conducted.

The internal numeric ID of the stage guarantees a correct
ordering (assuming, of course, that the stages were captured in the
correct order!).  Since only the relative ordering between the stages is
significant, the internal value is not exposed to clients; rather, a new
integer value is assigned to give the *illusion* of a monotonically
increasing sequence ID for a stage.  Like all other resource
identifiers, only the leading alphanumeric component of the identifier
is actually significant, with the trailing components providing
human-grokkable context.