Add new stages using the WS API.
Stages must be put using a token.  Attempting to add a stage with in an
invalid token will result in a 409 (Conflict).  The token reflects the
state of the database at the time of a `get` call and must match the
state of the DB at the time of the `put` call for it to be successful.

This puts an additional burden on the client to ensure that its view of
the stages is up to date and risks starvation if the mutation rate of
the stages is high.  This is still better, however, than having spurious
data being written to the DB.  Fortunately, the expected rate of change
of change is low, so in practice neither starvation nor the need for
frequent updates should be problematic.